access ProductDetailsModel in DetailsView at AppBar File
