+++ lib/core/utils/routes.dart
@@ -3,4 +3,5 @@ abstract class Routes {
   static String homeView = '/homeView';
   static const String loginView = '/loginView';
   static const String createAccountView = '/createAccountView';
+  static const String detailsView = '/detailsView';
 }
