+++ lib/features/home/presentation/views/widgets/dry_fruit_section_widget.dart
@@ -67,10 +67,7 @@ class DryFruitSectionWidget extends StatelessWidget {
                             )
                           : state is GetDryFruitProductSuccess
                               ? CustomProductWidget(
-                                  name: state.productDetails[index].name,
-                                  image: state.productDetails[index].image,
-                                  price: state.productDetails[index].price,
-                                  rate: state.productDetails[index].rate,
+                                  product: state.productDetails[index],
                                 )
                               : const Center(
                                   child: Text(
